+++ src/main/java/com/kurlabo/backend/repository/InstaSrcRepository.java
@@ -1,4 +1,4 @@
-package com.kurlabo.backend.repository.db;
+package com.kurlabo.backend.repository;
 
 import com.kurlabo.backend.model.db.Insta_src;
 import org.springframework.data.jpa.repository.JpaRepository;
