+++ src/main/java/com/kurlabo/backend/repository/SlideImgRepository.java
